Framework V2. Separating logic between pages.
