+++ client/game.html
@@ -6,6 +6,7 @@
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self'">
     <title>Hello World!</title>
     <link rel="stylesheet" href="./public/style/main.css">
+    <script src="./public/logic/gameLogic.js"></script>
   </head>
   <body>
     <audio disablewebsecurity crossorigin id="primaryAudio" src="./public/sounds/sound.wav"></audio>
