+++ client/menu.html
@@ -1,19 +1,22 @@
 <!DOCTYPE html>
 <html>
-<head>
-    <meta charset="UTF-8">
-    <title>Hello Page</title>
-    <style>
-        body {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            margin: 0;
-        }
-    </style>
-</head>
-<body>
-<h1>Hello</h1>
-</body>
+    <head>
+        <meta charset="UTF-8">
+        <title>Hello Page</title>
+        <style>
+            body {
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                height: 100vh;
+                margin: 0;
+            }
+        </style>
+        <script src="./public/logic/mainMenuLogic.js"></script>
+    </head>
+    <body>
+        <h1>Hello! This is the main menu.</h1>
+        <button id="backButton">Back to Game</button>
+    </body>
+
 </html>
